- fix: sql scheme
- feat: add ability to check that profile is member of specific channel

+++ db/migration/000001_init_profile.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS profile (
     id SERIAL PRIMARY KEY,
-    telegram_id INT UNIQUE,
-    telegram_chat_id INT,
+    telegram_id BIGINT UNIQUE,
+    telegram_chat_id BIGINT,
     username TEXT,
     preferred_currency VARCHAR(127),
     preferred_language VARCHAR(127),
